# Allocate parent node data alongside the array of children