I recently came across the [slice-dst](https://docs.rs/slice-dst/1.5.1/slice_dst/) crate, which lets you heap-allocate a struct adjacent to an array with a dynamic number of items. I realized that this kind of memory layout would probably be a good way to optimize in the core Tree-sitter library.

It fits in pretty easily. The code path for allocating parent nodes was already based on taking ownership of an array of child nodes. Now, instead of performing a separate `malloc` call for the parent data, I just grow the child node buffer large enough to hold the parent data, and add the parent data at the end.

This is valid because the children array is already pointer-aligned, so it has at *least* as large an alignment size as the parent data struct.

### Results

The change improves both memory usage and overall speed.

Here is a comparison of the memory used when parsing `jquery.js` (a 242k file):

|  | memory usage |
|-|---------|
| before | 4543.93 kB |
| after | 4127.50 kB |
| **% improvement**  | **9%** |

Here is a comparison of the speed of *parsing* and *syntax highlighting* all of the JavaScript files in the [`npm/cli`](https://github.com/npm/cli) (2614 files):

|  | parsing | highlighting |
|-|---------|------------|
| before | 1.83 seconds | 3.36 seconds |
| after | 1.56 seconds | 3.10 seconds |
| **% improvement**  | **15%** | **8%** |

In these flame graphs, I've put boxes around the allocation calls, and `ts_tree_delete`, which frees an entire syntax tree. In the second graph, one of the allocation code paths is eliminated, and the deletion code path is much faster.

**Before:**
<img width="1672" alt="alloc-old" src="https://user-images.githubusercontent.com/326587/97053085-1125b300-1537-11eb-93c0-08c25531a206.png">

**After:**
<img width="1671" alt="alloc-new" src="https://user-images.githubusercontent.com/326587/97053080-0ec35900-1537-11eb-8f31-0d22864d32be.png">